requirements.txt
Flask~=2.0.1
python-dotenv~=0.17.1
requests==2.21.0
pymongo~=3.11.4
dnspython~=2.1.0
pillow~=8.2.0
gunicorn~=20.1.0